Extension rewrite (#294)
* add test

* Hooking up manifestinfo file and

* moving files around and moving to a real yaml parser

* fixing build

* fixing build

* pulling up details to FEatureInfo

* First test on ExtensionInfo green whoop

* Adding in MAnifest Builder, now you can have multiple manifests for a module.

* fixing compoilation

* passing tests

* adding priority using a double

* movign around name spaces and addign feature manager

* Extensions Info List with ExtensionLoaders

* First hit at hookup

* Starting work on features

* Adding ShellFeaturesManager

* Adding Enable and Disable Features

* Adding Disable Features

* Define manifest file only.... Manifest defines what the extension is.

* Define manifest file higher up

* Adding FeatureDependencyNotification back in

* compiling

* app starts up

* bug fixes

* Setup up an running

* Fixing tests

* Changign casing of tests as mac is case sensitive with filenames

* Adding IFeatureInfoList

* fixing build

* reacting to PR comments

* tweaks

* Fixing build

* quicker initialize

* Fixing compilation issues after rebase

* Changed some stuff in the empty info lists

* site kicked up

* Add theme subpath

* more work to get setup working again

* some cleanup

* change everything to use Id and no, Id or Name

* more name to id changes

* Retrun Theme Extension Info

* fixing rebase

* IsModule check

* FeatureHash to run form Id no Name

* cleanup of not found stuff

* more cleanup

* more id work

* more id work

* more id work

* more id work

* more id work

* more id work

* more id work

* bit of perf

* Fixing extension paths

* Check fo is exists on directory info

* More changes for modules screen

* fixing yaml format

* fixing build

* fixing Features

* fixing json file

* looking at api confusion

* Fixing not awaited async calls

* Make all things async once more

* Updating ShellState to async

* React to PR comments

* more reacts

* Async changes and also duplicate strings in feature manager

* forgot to save some files

* clean up

* use id

* fixing build

* remove any clause to move away from link and use Count

* use interface

* removing case insensitive match on filename as will fail on mac

* Dont need to load features to show whats available, just use Features

* fixing json file

* react to PR

* React to PR comment, jsut return task

+++ src/Orchard.Environment.Extensions.Abstractions/Utility/DependencyOrderingUtility.cs
@@ -17,7 +17,7 @@ namespace Orchard.Environment.Extensions.Utility
         /// is a dependency of another, this algorithm will return the collection of elements sorted
         /// so that a given element in the sequence doesn't depend on any other element further in the sequence.
         /// </summary>
-        public static IEnumerable<T> OrderByDependenciesAndPriorities<T>(this IEnumerable<T> elements, Func<T, T, bool> hasDependency, Func<T, int> getPriority)
+        public static IEnumerable<T> OrderByDependenciesAndPriorities<T>(this IEnumerable<T> elements, Func<T, T, bool> hasDependency, Func<T, double> getPriority)
         {
             var population = elements.Select(d => new Linkage<T>
             {
@@ -40,7 +40,7 @@ namespace Orchard.Environment.Extensions.Utility
             return result;
         }
 
-        private static void Add<T>(Linkage<T> item, ICollection<T> list, IEnumerable<Linkage<T>> population, Func<T, T, bool> hasDependency, Func<T, int> getPriority)
+        private static void Add<T>(Linkage<T> item, ICollection<T> list, IEnumerable<Linkage<T>> population, Func<T, T, bool> hasDependency, Func<T, double> getPriority)
         {
             if (item.Used)
                 return;
@@ -55,9 +55,9 @@ namespace Orchard.Environment.Extensions.Utility
             list.Add(item.Element);
         }
 
-        private static int BestPriorityPosition<T>(List<T> list, int index, Func<T, T, bool> hasDependency, Func<T, int> getPriority)
+        private static int BestPriorityPosition<T>(List<T> list, int index, Func<T, T, bool> hasDependency, Func<T, double> getPriority)
         {
-            int bestPriority = getPriority(list[index]);
+            double bestPriority = getPriority(list[index]);
             int bestIndex = index;
 
             for (int i = index - 1; i >= 0; i--)
@@ -67,7 +67,7 @@ namespace Orchard.Environment.Extensions.Utility
                     return bestIndex;
                 }
 
-                int currentPriority = getPriority(list[i]);
+                double currentPriority = getPriority(list[i]);
                 if (currentPriority > bestPriority)
                 {
                     bestIndex = i;
